Process data as the shiny application

A utility function for reading a processing the data as `tagtango' does.

Usage

process_data(
  filename,
  data_type,
  left,
  right,
  pc1_axis1 = NULL,
  pc1_axis2 = NULL,
  pc2_axis1 = NULL,
  pc2_axis2 = NULL,
  filter_variable = NULL,
  filter_values = NULL,
  grouping_variable = NULL,
  grouping_values = NULL,
  min_counts = NULL,
  dimension = NULL,
  input_data = NULL
)

Arguments

filename

path to data file

data_type

experiment used in the MultiAssayExperiment or data contained in logcounts in SingleCellExperiment

left

annotation in the left of the diagram

right

annotation in the right of the diagram

pc1_axis1

first axis of first dimension reduction space

pc1_axis2

second axis of first dimension reduction space

pc2_axis1

first axis of second dimension reduction space

pc2_axis2

second axis of second dimension reduction space

filter_variable

filtering variable found in `dat'

filter_values

values in `filter_variable' that need to be excluded

grouping_variable

grouping variable found in `dat'

grouping_values

values in `grouping_variable' that need to be filtered

min_counts

minimum number of cells in a link for this to be displayed

dimension

internal variable to determine whether the data is a data frame (=0), a low-dimension dataset (=1), or a high-dimension dataset (=2).

input_data

object inputed directly.

Value

Returns a list containing two attributes: data and network. The first contains the data filtered according to the grouping and filtering values. The second, contains the data organized in a manner that is readible by the sankeyNetwork.