Generate a ggplot rose plot with marker expression

A function to generate a rose plot as `tagtango' does.

Usage

rose_plot(
  norm,
  data,
  selected,
  n_petals = 10,
  title = NULL,
  valley = NULL,
  palette = "RdYlGn",
  quant = c(1, 6),
  ...
)

Arguments

norm

normalized expression data with cells as rows and columns as markers/genes (i.e. function `process_data()' should provide this under the attribute `data')

data

data.frame with the different annotations, where each row represent a cell. The rownames should match those of `norm' (i.e. function `process_data()' should provide this under the attribute `network').

selected

boolean array with selected cells as TRUE values, for all cells in `data'.

n_petals

if an integer value, it defines the number of petals of the rose plot, selecting those that are most "relevant". If an array with marker names, it uses those.

title

main plot title. Default `NULL'.

valley

the value in `norm' corresponding to the valley separating positive and negative peak for CITE-seq data.

palette

color palette, default "RdYlGn".

quant

position of positive and negative peak for normalized data.

...

parameters passed to the underlying function.

Value

returns a ggplot object.